requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
psycopg2-binary
sqlalchemy
playwright
beautifulsoup4
requests
streamlit
python-dotenv
anthropic
